refactor: Cleanup polaris package
