+++ pkg/configauditreport/doc.go
@@ -1,0 +1,2 @@
+// This package provides primitives for working with Kubernetes workload configuration checkers.
+package configauditreport
