+++ pkg/configauditreport/plugin.go
@@ -16,5 +16,7 @@ import (
 type Plugin interface {
 	GetScanJobSpec(workload kube.Object, gvk schema.GroupVersionKind) (corev1.PodSpec, error)
 
+	GetContainerName() string
+
 	ParseConfigAuditResult(logsReader io.ReadCloser) (v1alpha1.ConfigAuditResult, error)
 }
