+++ pkg/configauditreport/scanner_test.go
@@ -1,0 +1,1 @@
+package configauditreport_test
