+++ pkg/configauditreport/writer.go
@@ -3,28 +3,27 @@ package configauditreport
 import (
 	"context"
 
-	clientset "github.com/aquasecurity/starboard/pkg/generated/clientset/versioned"
+	"github.com/aquasecurity/starboard/pkg/apis/aquasecurity/v1alpha1"
+	"github.com/aquasecurity/starboard/pkg/generated/clientset/versioned"
+	"github.com/aquasecurity/starboard/pkg/kube"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/klog"
-
-	starboard "github.com/aquasecurity/starboard/pkg/apis/aquasecurity/v1alpha1"
-	"github.com/aquasecurity/starboard/pkg/kube"
 )
 
 // Write is the interface that wraps basic methods for persisting ConfigAudit reports.
 //
 // Write persists the given ConfigAuditReport report.
 type Writer interface {
-	Write(ctx context.Context, report starboard.ConfigAuditReport) error
+	Write(ctx context.Context, report v1alpha1.ConfigAuditReport) error
 }
 
 // Reader is the interface that wraps basic methods for reading ConfigAudit reports.
 //
 // Read will return a single ConfigAuditReport that match a specific workload
 type Reader interface {
-	FindByOwner(ctx context.Context, owner kube.Object) (*starboard.ConfigAuditReport, error)
+	FindByOwner(ctx context.Context, owner kube.Object) (*v1alpha1.ConfigAuditReport, error)
 }
 
 type ReadWriter interface {
@@ -33,17 +32,17 @@ type ReadWriter interface {
 }
 
 type readWriter struct {
-	client clientset.Interface
+	clientset versioned.Interface
 }
 
-func NewReadWriter(client clientset.Interface) ReadWriter {
+func NewReadWriter(clientset versioned.Interface) ReadWriter {
 	return &readWriter{
-		client: client,
+		clientset: clientset,
 	}
 }
 
-func (w *readWriter) Write(ctx context.Context, report starboard.ConfigAuditReport) error {
-	existing, err := w.client.AquasecurityV1alpha1().ConfigAuditReports(report.Namespace).
+func (w *readWriter) Write(ctx context.Context, report v1alpha1.ConfigAuditReport) error {
+	existing, err := w.clientset.AquasecurityV1alpha1().ConfigAuditReports(report.Namespace).
 		Get(ctx, report.Name, metav1.GetOptions{})
 
 	if err == nil {
@@ -52,14 +51,14 @@ func (w *readWriter) Write(ctx context.Context, report starboard.ConfigAuditRepo
 		deepCopy.Labels = report.Labels
 		deepCopy.Report = report.Report
 
-		_, err = w.client.AquasecurityV1alpha1().ConfigAuditReports(report.Namespace).
+		_, err = w.clientset.AquasecurityV1alpha1().ConfigAuditReports(report.Namespace).
 			Update(ctx, deepCopy, metav1.UpdateOptions{})
 		return err
 	}
 
 	if errors.IsNotFound(err) {
 		klog.V(3).Infof("Creating ConfigAuditReport %q", report.Namespace+"/"+report.Name)
-		_, err = w.client.AquasecurityV1alpha1().ConfigAuditReports(report.Namespace).
+		_, err = w.clientset.AquasecurityV1alpha1().ConfigAuditReports(report.Namespace).
 			Create(ctx, &report, metav1.CreateOptions{})
 		return err
 	}
@@ -67,8 +66,8 @@ func (w *readWriter) Write(ctx context.Context, report starboard.ConfigAuditRepo
 	return err
 }
 
-func (w *readWriter) FindByOwner(ctx context.Context, workload kube.Object) (*starboard.ConfigAuditReport, error) {
-	list, err := w.client.AquasecurityV1alpha1().ConfigAuditReports(workload.Namespace).List(ctx, metav1.ListOptions{
+func (w *readWriter) FindByOwner(ctx context.Context, workload kube.Object) (*v1alpha1.ConfigAuditReport, error) {
+	list, err := w.clientset.AquasecurityV1alpha1().ConfigAuditReports(workload.Namespace).List(ctx, metav1.ListOptions{
 		LabelSelector: labels.Set{
 			kube.LabelResourceKind:      string(workload.Kind),
 			kube.LabelResourceName:      workload.Name,
