+++ pkg/polaris/doc.go
@@ -1,0 +1,2 @@
+// This package provides primitives for working with Polaris.
+package polaris
