+++ pkg/polaris/plugin_test.go
@@ -1,8 +1,12 @@
 package polaris_test
 
 import (
+	"os"
 	"testing"
+	"time"
 
+	"github.com/aquasecurity/starboard/pkg/apis/aquasecurity/v1alpha1"
+	"github.com/aquasecurity/starboard/pkg/ext"
 	"github.com/aquasecurity/starboard/pkg/kube"
 	"github.com/aquasecurity/starboard/pkg/polaris"
 	"github.com/aquasecurity/starboard/pkg/starboard"
@@ -10,8 +14,14 @@ import (
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/utils/pointer"
+)
+
+var (
+	fixedTime  = time.Now()
+	fixedClock = ext.NewFixedClock(fixedTime)
 )
 
 func TestPlugin_GetScanJobSpec(t *testing.T) {
@@ -93,11 +103,74 @@ func TestPlugin_GetScanJobSpec(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			plugin := polaris.NewPlugin(tc.config)
+			plugin := polaris.NewPlugin(fixedClock, tc.config)
 			jobSpec, err := plugin.GetScanJobSpec(tc.workload, tc.gvk)
 			require.NoError(t, err, tc.name)
 			assert.Equal(t, tc.expectedJobSpec, jobSpec, tc.name)
 		})
 	}
 
+}
+
+func TestPlugin_GetContainerName(t *testing.T) {
+	plugin := polaris.NewPlugin(fixedClock, starboard.ConfigData{})
+	assert.Equal(t, "polaris", plugin.GetContainerName())
+}
+
+func TestPlugin_ParseConfigAuditResult(t *testing.T) {
+	file, err := os.Open("testdata/polaris-report.json")
+	require.NoError(t, err)
+	defer func() {
+		_ = file.Close()
+	}()
+
+	config := starboard.ConfigData{
+		"polaris.imageRef": "quay.io/fairwinds/polaris:3.0",
+	}
+	plugin := polaris.NewPlugin(fixedClock, config)
+	result, err := plugin.ParseConfigAuditResult(file)
+	require.NoError(t, err)
+	assert.Equal(t, v1alpha1.ConfigAuditResult{
+
+		UpdateTimestamp: metav1.NewTime(fixedTime),
+		Scanner: v1alpha1.Scanner{
+			Name:    "Polaris",
+			Vendor:  "Fairwinds Ops",
+			Version: "3.0",
+		},
+		PodChecks: []v1alpha1.Check{
+			{
+				ID:       "hostIPCSet",
+				Message:  "Host IPC is not configured",
+				Success:  true,
+				Severity: "error",
+				Category: "Security",
+			},
+			{
+				ID:       "hostNetworkSet",
+				Message:  "Host network is not configured",
+				Success:  true,
+				Severity: "warning",
+				Category: "Networking",
+			},
+		},
+		ContainerChecks: map[string][]v1alpha1.Check{
+			"db": {
+				{
+					ID:       "cpuLimitsMissing",
+					Message:  "CPU limits are set",
+					Success:  true,
+					Severity: "warning",
+					Category: "Resources",
+				},
+				{
+					ID:       "cpuRequestsMissing",
+					Message:  "CPU requests are set",
+					Success:  true,
+					Severity: "warning",
+					Category: "Resources",
+				},
+			},
+		},
+	}, result)
 }
